#7: Moved useinterval into separate file
