+++ components/ColorPad/index.tsx
@@ -1,27 +1,10 @@
 import { View, Text, StyleSheet } from "react-native";
 import { useEffect,useRef,useState } from "react";
 import * as Animatable from 'react-native-animatable';
+import {useInterval} from "../../helpers";
 import { Color, Screen, generateColor,Time,GameMode } from "../../constants";
 import { usePrevious } from "../../hooks";
 
-
-const useInterval = (callback, delay) => {
-  const savedCallback = useRef<Function>();
-
-  useEffect(() => {
-    savedCallback.current = callback;
-  }, [callback]);
-
-  useEffect(() => {
-    const tick = () => {
-      savedCallback.current();
-    }
-    if (delay !== null) {
-      let id = setInterval(tick, delay);
-      return () => clearInterval(id);
-    }
-  }, [delay]);
-};
